[Integrate iot app with facerognitioni & facetracker] - fix integration bugs. - delete build folder - add Nodered nodes diagram

+++ IoT/iot-android/facerecognition/src/main/java/com/microsoft/projectoxford/face/samples/persongroupmanagement/PersonActivity.java
@@ -262,13 +262,18 @@ public class PersonActivity extends AppCompatActivity {
 
             @Override
             public boolean onActionItemClicked(ActionMode mode, MenuItem item) {
-                switch (item.getItemId()) {
-                    case R.id.menu_delete_items:
-                        deleteSelectedItems();
-                        return true;
-                    default:
-                        return false;
+                if(item.getItemId() == R.id.menu_delete_items) {
+                    deleteSelectedItems();
+                    return true;
                 }
+                return false;
+//                switch (item.getItemId()) {
+//                    case R.id.menu_delete_items:
+//                        deleteSelectedItems();
+//                        return true;
+//                    default:
+//                        return false;
+//                }
             }
 
             @Override
